+++ src/mml_utils/phenorm/afep.py
@@ -2,7 +2,13 @@
 Implement AFEP, including Greedy Feature Selection.
 """
 import json
+import datetime
+import math
+import pathlib
 
+import pandas as pd
+
+from mml_utils.phenorm.cui_expansion import add_shorter_match_cuis
 from loguru import logger
 
 from mml_utils.parse.json import extract_mml_from_json_data
@@ -79,3 +85,55 @@ def run_greedy_algorithm(cui_df, df):
         temp_df = temp_df[temp_df[cui] == 0.0]
     logger.info(f'Result of greedy algorithm: retained {len(res)} CUIs.')
     return res
+
+
+def run_afep_algorithm(note_directories, *, mml_format='json', outdir: pathlib.Path = None,
+                       expand_cuis=False, apikey=None, skip_greedy_algorithm=False, min_kb=None, data_directory=None):
+    now = datetime.datetime.now().strftime('%Y%m%d_%H%M%S')
+    if outdir:
+        outdir.mkdir(exist_ok=True, parents=True)
+    else:
+        outdir = pathlib.Path('.')
+
+    article_types, results = extract_articles(note_directories, mml_format, data_directories=data_directory)
+    logger.info(f'Article types: {article_types}')
+
+    if expand_cuis:
+        results = add_shorter_match_cuis(results, apikey)
+
+    logger.info(f'Building pandas dataset from {len(results)} results.')
+    df = pd.DataFrame.from_records(results)
+    s = df[['cui', 'article_source']].drop_duplicates().groupby('cui').count()
+    if min_kb is None:
+        min_kb = math.ceil(len(article_types) / 2)
+    logger.info(f'Retaining only CUIs appearing in at least {min_kb} knowledge base sources.')
+    cuis_with_three_or_more = set(s[s.article_source >= min_kb].index)
+    cui_df = df[df.cui.isin(cuis_with_three_or_more)].copy()
+    logger.info(f'Retained {cui_df.shape[0]} CUIs.')
+
+    # greedy algorithm to reduce number of CUIs
+    if skip_greedy_algorithm:
+        logger.info(f'Skipping greedy algorithm.')
+        res = cui_df['cui'].unique()
+    else:
+        res = run_greedy_algorithm(cui_df, df)
+
+    res_dict_df = cui_df[cui_df['cui'].isin(res)].groupby(['cui', 'preferredname']).agg({
+        'conceptstring': lambda x: ','.join(set(x)),
+        'matchedtext': lambda x: ','.join(set(x)),
+        'all_sources': lambda x: ','.join(set(x)),
+        'all_semantictypes': lambda x: ','.join(set(x)),
+    }).reset_index()
+
+    with open(outdir / f'selected_cuis_{now}.txt', 'w') as out:
+        out.write('\n'.join(res_dict_df['cui'].unique()))
+
+    cui_df['value'] = 1
+    article_df = cui_df[['cui', 'article_source', 'value']].pivot_table(
+        index='cui', columns='article_source', fill_value=0
+    )
+    article_df.columns = article_df.columns.droplevel(level=0)
+    article_df.columns.name = None
+    article_df['n_articles'] = article_df.apply(sum, axis=1)
+    summary_df = pd.merge(res_dict_df, article_df.reset_index(), on='cui')
+    summary_df.to_csv(outdir / f'selected_cui_details_{now}.csv', index=False)